BAVL-235 creating dedicated integration base class for localstack and turning off on CI pipline to avoid potential timing issues affecting builds.
