+++ src/test/resources/application-test-localstack.yml
@@ -34,3 +34,19 @@ api:
       prisoner-search: http://localhost:8092
       manage-users: http://localhost:8093
       prison-api:  http://localhost:8094
+
+hmpps.sqs:
+  provider: localstack
+  queues:
+    bvls:
+      queueName: ${random.uuid}
+      dlqName: ${random.uuid}
+      subscribeTopicId: domainevents
+  topics:
+    domainevents:
+      arn: arn:aws:sns:eu-west-2:000000000000:${random.uuid}
+
+feature:
+  events:
+    sns:
+      enabled: true
